[#54] refactor(*): gatsby v3
