+++ package.json
@@ -8,62 +8,61 @@
     "url": "https://github.com/hideyuk1-jp/beyooon.jp/issues"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.5",
-    "@emotion/styled": "^10.0.27",
-    "@types/react-helmet": "^6.1.0",
-    "@types/yup": "^0.29.10",
-    "axios": "^0.21.0",
-    "emotion-theming": "^10.0.27",
-    "formik": "^2.2.5",
-    "gatsby": "^2.25.1",
-    "gatsby-image": "^2.4.21",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.3.0",
+    "@types/react-helmet": "^6.1.1",
+    "@types/yup": "^0.29.11",
+    "axios": "^0.21.1",
+    "formik": "^2.2.6",
+    "gatsby": "^3.4.0",
+    "gatsby-image": "^3.4.0",
     "gatsby-plugin-draft": "^0.1.0",
-    "gatsby-plugin-emotion": "^4.4.0",
-    "gatsby-plugin-feed": "^2.6.0",
+    "gatsby-plugin-emotion": "^6.4.0",
+    "gatsby-plugin-feed": "^3.4.0",
     "gatsby-plugin-google-adsense": "^1.1.3",
-    "gatsby-plugin-google-analytics": "^2.4.0",
-    "gatsby-plugin-manifest": "^2.5.1",
-    "gatsby-plugin-offline": "^3.3.1",
-    "gatsby-plugin-react-helmet": "^3.3.14",
-    "gatsby-plugin-sharp": "^2.7.0",
-    "gatsby-plugin-sitemap": "^2.8.0",
-    "gatsby-plugin-typegen": "^2.2.1",
+    "gatsby-plugin-google-analytics": "^3.4.0",
+    "gatsby-plugin-manifest": "^3.4.0",
+    "gatsby-plugin-offline": "^4.4.0",
+    "gatsby-plugin-react-helmet": "^4.4.0",
+    "gatsby-plugin-sharp": "^3.4.0",
+    "gatsby-plugin-sitemap": "^4.0.0",
+    "gatsby-plugin-typegen": "^2.2.4",
     "gatsby-remark-code-titles": "^1.1.0",
-    "gatsby-remark-copy-linked-files": "^2.3.19",
-    "gatsby-remark-images": "^3.4.0",
-    "gatsby-remark-prismjs": "^3.6.0",
-    "gatsby-remark-responsive-iframe": "^2.4.17",
-    "gatsby-remark-smartypants": "^2.3.13",
-    "gatsby-source-filesystem": "^2.4.0",
-    "gatsby-transformer-remark": "^2.9.1",
-    "gatsby-transformer-sharp": "^2.5.21",
+    "gatsby-remark-copy-linked-files": "^4.1.0",
+    "gatsby-remark-images": "^5.1.0",
+    "gatsby-remark-prismjs": "^5.1.0",
+    "gatsby-remark-responsive-iframe": "^4.1.0",
+    "gatsby-remark-smartypants": "^4.1.0",
+    "gatsby-source-filesystem": "^3.4.0",
+    "gatsby-transformer-remark": "^4.1.0",
+    "gatsby-transformer-sharp": "^3.4.0",
     "moment": "^2.29.1",
-    "moment-timezone": "^0.5.32",
-    "prismjs": "^1.22.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "moment-timezone": "^0.5.33",
+    "prismjs": "^1.23.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-emoji-render": "^1.2.4",
     "react-helmet": "^6.1.0",
-    "react-share": "^4.3.1",
-    "yup": "^0.32.1"
+    "react-share": "^4.4.0",
+    "yup": "^0.32.9"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.0.1",
-    "@typescript-eslint/parser": "^4.0.1",
-    "eslint": "^7.12.1",
-    "eslint-config-prettier": "^6.15.0",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
+    "eslint": "^7.25.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-config-react-app": "^6.0.0",
     "eslint-loader": "^4.0.2",
-    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint-plugin-flowtype": "^5.7.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "gatsby-plugin-eslint": "^2.0.8",
-    "gatsby-plugin-typescript": "^2.5.0",
+    "gatsby-plugin-eslint": "^3.0.0",
+    "gatsby-plugin-typescript": "^3.4.0",
     "prettier": "^2.1.2",
-    "typescript": "^4.0.5"
+    "typescript": "^4.2.4"
   },
   "homepage": "https://github.com/hideyuk1-jp/beyooon.jp",
   "keywords": [
