+++ src/components/atoms/time-to-read.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styled from '../atoms/styled';
+import styled from '@emotion/styled';
 
 type Size = 'small' | 'medium';
