+++ src/components/organisms/header/hamburger.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styled from '../../atoms/styled';
+import styled from '@emotion/styled';
 
 const Hamburger: React.FC<{
   isOpen: boolean;
