+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import { PageProps } from 'gatsby';
 
-import styled from '../components/atoms/styled';
+import styled from '@emotion/styled';
 import Layout from '../components/templates/layout';
 import SEO from '../components/organisms/seo';
 import Hero from '../components/organisms/hero';
