+++ src/styles/global.ts
@@ -1,4 +1,4 @@
-import { css } from '@emotion/core';
+import { css } from '@emotion/react';
 
 const resetStyle = css`
   /* Box sizing rules */
