+++ src/styles/syntax-highlight.ts
@@ -1,4 +1,4 @@
-import { css } from '@emotion/core';
+import { css } from '@emotion/react';
 
 const SytaxHighlightStyle = css`
   /**
