+++ src/templates/blog-post.tsx
@@ -8,7 +8,7 @@ import Layout from '../components/templates/layout';
 import SEO from '../components/organisms/seo';
 import ShareButtons from '../components/organisms/share-buttons';
 
-import styled from '../components/atoms/styled';
+import styled from '@emotion/styled';
 import SyntaxHighlightStyle from '../styles/syntax-highlight';
 import TimeToRead from '../components/atoms/time-to-read';
 import PostDate from '../components/atoms/post-date';
