+++ src/types/emotion.d.ts
@@ -1,0 +1,6 @@
+import '@emotion/react';
+import { Theme as MyTheme } from './index';
+
+declare module '@emotion/react' {
+  export interface Theme extends MyTheme {}
+}
